Remove legacy headers from new dev API
